Done with VIF experiments
I've been chasing VIFs for a while now without much success. The problem that I've discovered is that many of regressors predict other regressors perfectly. For example, we have all the factors we need to calculate FICO such as revolving debt utilization, late payments, and delinquent accounts. The only thing that would be included in FICO that isn't in the data is public records such as a tax lien or civil judgement. I'm guessing people with those types of credit hits are filtered out by lending club.

I dropped a lot of redundant data and the scores are only improving marginally.